Update integrations-core pin and add FIPS workaround for psycopg (#41003)
### What does this PR do?

Updates integrations-core to a version using a custom-built version of psycopg. It also adds a matching workaround for the FIPS version of the Agent following what was previously done for the `cryptography` library.

### Motivation

#incident-43132

Problems with no compatible cyphers being found for the postgres integration on FIPS versions of the Agent. These surfaced after updating from psycopg2 to 3, which bumped openssl from 1 to 3.

### Describe how you validated your changes

We deployed these changes to staging for validation.

### Additional Notes


Co-authored-by: alopezz <alex.lopez@datadoghq.com>