CONTINT-4803 - expose provider-kind as static tag (#40944)
### What does this PR do?

the tag `provider-kind` used to be exposed as GCE specific tag.

The tag is static, it's set once on cluster creation, then will never change. it's strongly linked to how the cluster was created.

For that, read the env value `DD_PROVIDER_KIND` that exposes if the provider kind and set the tag once when the agent starts.

That static tag list used to be only for fargate hosts, now it's for GKE too, adjust the logic to make it provider independent.

Remove the code and the test in GCE specific tag logic.

closes: CONTINT-4803

### Motivation

See CONTINT-4803

### Describe how you validated your changes

Added unit-tests

### Additional Notes

manually set the env variable `DD_PROVIDER_KIND` and check the UI on a random metric for the tag.