dyninst/*: overhaul module, actuator, and interfaces  (#41102)
### What does this PR do?

This change seeks to reduce the conceptual overhead of the dyninst
subsystem. In particular, it decouples the data plane and probe
generation internals from the actuator package. The actuator now really
is just a coordinator of actions that are encapsulated by a `Runtime`. A
`Runtime` can load and attach programs.

It is the responsibility of the runtime to hook up and manage the the
data its probes produce.

This change also eliminates the hard-to-explain `module.Controller`
type, which was hard to differentiate for the `module.Module` type. The
`module.Module` now also has much more explicit dependencies.

The end result feels cleaner, and will allow us to evolve the way
scraping works without so much coupling to concepts like the `ir`. That
coupling was accidental.

In the short term, it does mean there's a bit of duplication in the
`module` and `rcscrape` packages, but I think that's better than the
coupling they had before.

### Motivation

The abstractions we grew as we built dyninst needed a revision. We found it painful to
decouple scraping from the ir and the eBPF program we use for everything else while
still benefiting from the coordination provided by the actuator. This fixes that problem.

Along the way I also added checks for goroutine leaks in our tests.

### Describe how you validated your changes

No tests were harmed in this change, and leak checking was added.

Fixes https://datadoghq.atlassian.net/browse/DEBUG-4199