[EBPF-810] update(ebpfcheck): use precise map mem usage when available (#41189)
### What does this PR do?

Here we update the memory usage estimation of ebpf hash maps for ebpf-checks, so that the `BPF_F_NO_PREALLOC` flag is properly handled. This is achieved by checking reading the fdinfo of maps from procfs, in which precise memory usage counting is exposed by recent kernel versions (6.4+). In all other cases, we fallback to the already existing method.

### Motivation

With ebpf maps with the `BPF_F_NO_PREALLOC` flag, using an heuristics based on max entries would be pessimistically misleading due to the fact that memory footprint of maps could grow at runtime.

### Describe how you validated your changes

For now, running locally the system-probe unit tests suite.

### Additional Notes