Fix raw data len calculation for JSON aggregation  (#41247)
### What does this PR do?

This PR fixes a bug when auto multiline detection is enabled, the JSON aggregator would record the logs byte size as `1` byte smaller than the actual offset in the file. This can cause the agent to re-collect a small number of bytes from previously tailed files when the agent restarts and resumes from the offset stored in the registry. 

### Motivation

Bug Fix

### Describe how you validated your changes

Unit test reproduces the issue, and validate the fix. 
Also tested manually by emitting a log file that is exactly 300 bytes, and observing the offset in the `registry.json` 

### Additional Notes