Compose `buildifier` from upstream and `multitool`
Currently, `buildifier` targets don't work out of the box on Windows -
credits to Ofek for reporting.

Sadly, it turns out that two separate attempts have already been made to
remedy this:
1. In the upstream `bazelbuild/buildtools` (`go` source, but `workspace`
   suitable):
   https://github.com/bazelbuild/buildtools/pull/1230 (merged)
2. `keith/buildifier-prebuilt` (`bzlmod` suitable, providing slightly
   outdated `go` binaries published by `bazelbuild/buildtools`):
   https://github.com/keith/buildifier-prebuilt/pull/89 (well tested,
   but pending CI feedback... for months!)

Inspired by Tony's
[tips](https://datadoghq.atlassian.net/browse/ABLD-174?focusedCommentId=2644283),
the present change proposes the "least intrusive" change so that:
- we don't need to start a third effort (rules, `.bash` and `.bat`
  scripts),
- we benefit from the validated and merged upstream behavior,
- we don't depend on the goodwill of `keith/buildifier-prebuilt` to
  publish the version `8.2.1` of the binaries (nor subsequent versions),
- we don't have to manage a complex patch, but instead simply inhibit
  the upstream lines that we do not need, given `buildifier`
  implementations are provided by `multitool` again.

In short, this change combines the best of both worlds through a simple
indirection made possible by the configurability of upstream rules
which, unlike `keith/buildifier-prebuilt`, allow us to define which
implementation of `buildifier` to use.