[EBPF-824] gpu: fix TestConsumerProcessExitViaCheckClosedProcesses flaky test (#41302)
### What does this PR do?

Fixes a low frequency flaky test. The way the test was written, the `procRoot` with the fake data was not being actually used to check for the closed processes (it was being passed to the context but not the consumer), and instead it was using the real `/proc` path. This made the test work most of the time, except when we had a real process with the PID of our fake process. 

### Motivation

Eliminate flaky tests.

### Describe how you validated your changes

Fixed the unit test.

### Additional Notes