Autoscaling: fix infinite update loop when spec hashes do not match (#41301)
### What does this PR do?

This PR fixes the behaviour of the Cluster Agent when the `.Spec` retrieved _after_ a DPA has been created/updated with values not originating from the Cluster Agent itself.

The discrepancy can come from different sources: admission controller, CRD defaulting, etc. which are not necessarily expected but should not break the overall feature either.

### Motivation

Fixing inactive remote Autoscalers when Remote Spec hash is different from actual object Spec hash after update.

### Describe how you validated your changes

Change validate in a testing Kubernetes cluster. It can be reproduced by creating a remote autoscaler with a recent version of the Cluster Agent/CRD that has defaulting for local fallback values.

Added a unit test to cover this specific case.

### Additional Notes

The current solution is not optimal and we cannot cope with differences manually done in case Cluster Agent leader was not available at the time of the update.