usm, cnm: Add process tags from tagger
Get the tags for the Process entity from the tagger and add them to the
connection tags.  If service discovery is enabled, this will include
things like UST (will be de-duplicated in the encoder since system-probe
also gets it directly for USM/CNM) as well as process tags from tracer
metadata.