[flakefinder] Fix pipeline generation with more than 50 jobs matrixes (#41298)
### What does this PR do?

The following error happened when generating the flake finder pipeline: https://gitlab.ddbuild.io/DataDog/datadog-agent/-/pipelines/77611250
It happened because new-e2e-windows-installer defines a matrix with more than 50 jobs. Adding a needs on that job is not possible because the limit is at max 50 needs.
To prevent that from happening again, just take the first element of the matrix if it would create a needs with more than 50 jobs. It changes a bit the need of the generated job, but the main goal is to make sure they do not run all at the same time, not because they really have a dependency between them

### Motivation

Fix flakefinder

### Describe how you validated your changes

### Additional Notes