dyninst/irgen: use Go ABI for return value locations
The Go calling convention and ABI is well-defined regarding where return
values live in terms of registers and on the stack for out-of-line calls.
Using this information is crucial for return probes because the dwarf
information for these variables is so terrible.

This PR makes return probes actually capable of returning data.