dyninst: rework integration testing to use module (#41360)
### What does this PR do?

The integration tests were built before we had a clear
end-to-end abstraction. It's now bad that we aren't using it
and it was getting in the way of work to integrate return
probes.

This ends up being a pretty minimal change that expands test
coverage.

### Motivation

We'll need to use the whole of the module infrastructure and runtime in return probes. This is being pulled out of https://datadoghq.atlassian.net/browse/DEBUG-4031

### Describe how you validated your changes

It's just testing.