discovery: Do not re-walk fds for tracer metadata
The Service Discovery code walks the process' fd list once to get the
sockets and log files, and then (if it's a service) walks the list again
to find the memfd for the tracer metadata.  Eliminate the latter walk
and have the former walk note the location of the metadata memfd for
re-use by the metadata parsing code.