[Stackcleaner] Do not fail if issue when triggering stack delete, send event (#41386)
…

### What does this PR do?

Make sure the call to stack cleaner cannot fail the test.
It makes no sense to fail the test on clean up step.  To avoid flakiness we should ignore such error. An event is sent to datadog so that we can be aware of such issues

### Motivation

Prevent flakiness on steps that does not indicate anything about the test result

### Describe how you validated your changes

### Additional Notes