(fleet) disable config experiments on windows
This PR disables the config experiments on windows. With the recent changes that copy the configuration directory we failed to realise that on windows there is more than just the config in this directory and instead it is merged with runtime state and logs.

This PR fixes this by skipping the experiment process for configuration on windows.