dyninst/symdb: include some inlined functions in symdb output (#41292)
Before this patch, when iterating over the debug info in streaming mode,
we were ignoring all inline functions. Now, we include in the output
inline functions and methods that are inlined in their own compile unit
(which is the case for all unexported functions and methods), and we
also include stand-alone inline functions that are inlined in some other
compile unit -- we include these in the compile unit where their
abstract definition happens to appear in the DWARF.

For methods inlined in only in other compile units, we don't include
them in the output -- methods hang from types, and we don't want to
produce incomplete types or report the same type in multiple packages.