dyninst/*: add return probes (#41392)
### What does this PR do?

This PR adds return events for method snapshot probes. In so doing, it augments out-of-line method probes with duration, stack traces at the point of return, and return values (which are rarely available).

The basic design is to capture stack depth and to pair calls with returns by using a per-goroutine hash map in ebpf to track the depths of open calls. This needs to be bounded in terms of both number of goroutines and number of calls per goroutine.

Then in userspace we pair entry events with return events.

We use a marker byte to declare the disposition of an event with regards to finding its mate. 

We also augment the decoding logic to combine the events as the frontend expects.

### Describe how you validated your changes

There's some testing.

### Additional Notes

There is still future work:
 * Augmenting the dwarf information with return value availability as inferred from the ABI.
 * Handling recovery from panics and mapping that to returns.
 * Ensuring that stack hashes are made available for decoding even while events are buffered (this one may be a bug).

This is on top of https://github.com/DataDog/datadog-agent/pull/41368 and https://github.com/DataDog/datadog-agent/pull/41360.

https://datadoghq.atlassian.net/browse/DEBUG-4031