extend serializer pipeline model to sketches (#41366)
### What does this PR do?

Extends the serializer pipeline model to sketches, in addition to its current use for series data. This enables sketches to be send to the preaggregation endpoint in the same way as series.

### Motivation

We want to be able to send sketches to preaggregation.

### Describe how you validated your changes

Added unit tests and local verification with fakeintake.

### Additional Notes

Using the same pipelines for sketches as series also includes the local autoscaling feature, which was previously only applied to sketchs. This will not have any effect in practice, since that feature works off of an explicit allowlist (so a sketch would need to be configured) and we also have a filter for series payloads only at the point in the forwarder where we actually send requests.

TPOT-213