[SMP] Churn docker containers (#41098)
### What does this PR do?

Churn the docker containers used to load the agent in the `docker_containers_memory` and `docker_containers_cpu` SMP experiments.

### Motivation

There are functions in the agent that are triggered only as reaction to container events like creation and deletion. Auto-discovery for scheduling/un-scheduling checks for ex.
A static workload with no container creation and deletion cannot detect potential memory leaks in those kind of functions.

### Describe how you validated your changes

### Additional Notes

Uses DataDog/lading#1325.