(fleet) disable config experiments on windows (#41428)
This PR fixes a bug preventing the update of configurations through Fleet on windows.

With the recent switch to modifying the user configuration dir directly we failed to account that the "configuration" directory on windows also contains all the runtime files (python cache, installer state, etc...). This broke the current implementation of the experiment.

This PR fixes the issue by disabling config experiments on windows.

### QA

This change was QA'd manually.