[EBPF] gpu: skip cgroup tests when cgroupfs is readonly (#41506)
### What does this PR do?

Skips cgroup tests for `pkg/gpu` when the cgroupfs is not writable.

Additionally, it marks the oracle job as allowed to fail, as it started
to fail at the same point as the pkg/gpu tests.

### Motivation

FIx for #incident-43807

### Describe how you validated your changes

CI green. These tests also execute in KMT, where we have full access to
the cgroup, so we do not lose coverage.

### Additional Notes

(cherry picked from commit 46246af8bce3d52f9ff85550f72fac217cafd53a)