[corechecks/kubelet] Fix flaky test in pod provider (#41379)
### What does this PR do?

This PR fixes flaky tests in the pod provider of the kubelet check.

These tests were instantiating a new workloadmeta including its collectors on each test. This caused race conditions when multiple `kube_metadata` collectors ran at the same time.

To fix the flakyness,  I've extracted the functions that parse kubelet pods from the kubelet workloadmeta collector.

The kubelet check tests need to read a kubelet response from a file and fill workloadmeta with the pods and containers parsed. The parse logic was only inside the kubelet workloadmeta collector, which lives in an `internal` package. That forced us to instantiate workloadmeta with real collectors, which makes everything async and a bit complex. By extracting the parse function, tests are now way faster and easier to reason about.

### Describe how you validated your changes

CI