Stop sending the `example.metric` metric to the /sketches endpoint (#41495)
### What does this PR do?

We have a connectivity check that probes the /sketches endpoint--this is adding a new metric to a customer's account, so we want to not send this metric for customers. We send an empty payload, which still checks for connectivity, fulfilling the requirements of the check.

### Motivation

### Describe how you validated your changes

I built an agent with this changes, ran the `agent diagnose connectivity-check` command, and saw that the check was passing. If the payload was modified with miscellaneous data, the check fails, as expected.

### Additional Notes