Enhance BSOD reports with callstack and agent version (#41434)
### What does this PR do?
This PR enhances Windows BSOD reports to include the crash callstack and version of the Agent found in the dump.
The callstack does not have resolved symbols and should remain so per legal consultation.

Originally the driver version was intended to be reported as well, but the driver modules do not include embedded resource tables that have the version information. However, the Agent does have the resource table with version information and it can be queried from a system crash dump.
In any case, the Agent version is more meaningful than the driver version, as it better identifies the fuller end-to-end context (e.g. combination of features).

### Motivation
https://datadoghq.atlassian.net/browse/WINA-1869

### Describe how you validated your changes
Queried system-probe to report a prepared crash dump.
Utilized a VM with a prepared MEMORY.DMP, installed the agent, and verified the BSOD report was exercised. Verified that the notification was sent.
CI/CD, verified crash reporting related tests.