[RC] Support the new ECDSA based TUF chain for gov environments (#41529)
### What does this PR do?
Adds the open-source roots for the FIPS compliant ECDSA key based TUF roots. Due to nuances with how public keys are encoded within the TUF files, we cannot send canonicalized JSON down to clients. This was never needed in the first place though, since all verification that takes place internalizes canonicalization before validation of the signature.

### Motivation
Support RC in gov

### Describe how you validated your changes
Manually, by connecting the agent up to a test org in gov and verifying that TUF payloads sent by the backend were able to be validated by the agent

### Additional Notes