[EBPF-828] gpu: support retries of NVML init in system-probe (#41518)
### What does this PR do?

This PR changes the `DeviceCache` so that it's no longer initialized on creation, but rather on use. It requires changing the signature of the rest of `DeviceCache` methods to return an error. Each of them will check if the library is initialized, and if not they will try to do it. If the initialization fails, they will return an error.

It replicates the behavior we have in the core check, where we would be trying to initialize NVML libraries on each `Run()` until it initialized successfully. Now the system-probe module will start up even if NVML is not accessible, fixing issues where the driver might not have been loaded when system-probe started.

### Motivation

https://datadoghq.atlassian.net/browse/EBPF-828 

Ensuring compatibility of the system-probe module when using the NVIDIA GPU operator with operator-managed drivers, which can be loaded after system-probe starts.

### Describe how you validated your changes

Unit tests and e2e tests cover the refactor.

### Additional Notes