[EBPF] gpu: fix panic in stream handling (#41555)
### What does this PR do?

If we have an error getting the current stream, we should not try to process the event.

### Motivation

#incident-43873 

### Describe how you validated your changes

Added unit test reproducing the panic.

### Additional Notes

This failure path existed previously, where we would ignore the error from `getStream`, but in practice there were no situations were errors were emitted. However, with https://github.com/DataDog/datadog-agent/pull/41518 it was more frequent for this function to emit an error, which meant we were now triggering the bugged code path.