Pin agent-release-management to the 7.71.x branch (#41613)
### What does this PR do?

Updates the `7.71.x` branch of `datadog-agent` to use the `7.71.x`
branch of `agent-release-management`.
Pins the `dd-pkg` version to `0.50.0` to not include
https://github.com/DataDog/dd-pkg/pull/100, which prevents using
`dd-pkg` for prod releases if a non-`main` branch of
`agent-release-management` is used.

### Motivation

Pin `agent-release-management` for the `7.71.x` release to introduce new
changes on `main` without disrupting the ongoing release.

### Describe how you validated your changes

CI passes.

### Additional Notes

n/a