dyninst: fix plumbing of container ID to trace-agent (#41635)
Before this patch, we were only sending the container ID to the trace-agent while uploading logs when we also had other tags to send (i.e. the SCI tags). There was no reason for this; now we always send a container ID when we know one. The trace-agent needs the container ID to augment the logs with many other tags, like the service name.