Fix RemoteAgentRegistry Unit Test (#41687)
### What does this PR do?

This PR updates the TestGetRegisteredAgentsIdleTimeout unit test by introducing a short delay between the creation of the fake remoteAgent instance and its subsequent stop() call.

### Motivation

Without this delay, a race condition can occur between the non-blocking startup of the gRPC server and its shutdown. This causes the test to intermittently fail due to the server being stopped before it has fully started, leading to an error when attempting to start a gRPC server that has already been terminated.

### Describe how you validated your changes

I reran the test multiple times to ensure stability. However, since the issue is timing-dependent, reproducing it consistently is difficult. The added delay mitigates the observed race condition in repeated test runs.