fix(serverless-init): add nil channel check for service check and event flushing (#41539)
### What does this PR do?

Adds nil channel check for service check and event flushing.

### Motivation

https://datadoghq.atlassian.net/browse/SVLS-7697

When the Java tracer sends JVM metrics, it also sends a service check, `jmxfetch-config.can_connect`. The Serverless Batcher does not set the service check channel, so when the service checks are attempted to be flushed to a `nil` channel the goroutine is blocked and no more metrics can be sent through DogStatsD.

https://github.com/DataDog/datadog-agent/blob/0c5cd71e99e97b0f7f48b3841593e1f571cc7973/comp/dogstatsd/server/batch.go#L152-L177

### Describe how you validated your changes

- Added service check and event flush unit tests for both the Serverless Batcher and non Serverless Batcher.
- Deployed to Google Cloud Run and verified JVM metrics are sent to Datadog successfully
- Deployed to Azure Container Apps and verified JVM metrics are sent to Datadog successfully

### Additional Notes

The same issue with a nil service check channel also affects event channels, so added a nil check for events as well.