Inline installer build into datadog-agent build definition (#41620)
### What does this PR do?

Inlines the installer build into the datadog-agent Omnibus software definition, and merges both builds in Windows, where they were still separate.

### Motivation

Due to how Omnibus software definitions work, this was copying almost the entire datadog-agent repo twice for each build, as both installer.rb and datadog-agent.rb were using a path source to the root of the datadog-agent source code. This is not very noticeable in Linux builds, where these copies are happening in parallel, but it adds close to 10 minutes of build time to Windows builds, as the installer was being built in a separate omnibus run.

Rather than just merge builds in Windows to do the same as in Linux, it's better to also just avoid copying everything twice altogether. Even if parallelized, these would still unnecessarily slow down Windows builds down, knowing that IO performance is typically worse there.

### Describe how you validated your changes

Green CI.

### Additional Notes

Important note: This PR bumps the MSI on-disk quality gate because of how the change affects how the size is computed. We use the generated zip file, and not the MSI, when computing the uncompressed package size. Before this PR, the installer was not being built on the same zip as the regular Agent package, and thus the computation didn't include the installer's size. This can be confirmed by the fact that the size of the MSI file has barely changed.

I'm not sure whether we can now just remove the Omnibus installer project and copy, hence I'll probably defer that for a future follow-up.