[EBPF-763] attacher: fix flaky test (#41694)
### What does this PR do?

Fixes a flaky test by replicating the logic from other tests for process exiting. As the event stream might lose events, we check and re-send the event via a proxy so that we trigger the same code path, eliminating the event stream as a source of flakiness.

### Motivation

Reduce flakiness

### Describe how you validated your changes

Test is passing, checked multiple runs that it doesn't flake

### Additional Notes