Notify tracer metadata memfd creation via eBPF
There is code in the agent to get tags from the tracer metadata of
processes and attach them to network connections by using the tagger
component.  However, in the case of short-lived processes, or during
process startup, network data may be available before the process'
metadat is available in the agent. To significantly reduce the race
window, add an eBPF event to monitor the sealing of the memfd used by
the tracers, and read the memfd when that event is received.

We hook memfd_create(2) to check that the name of the memfd starts with
the prefix we want.  We store the fd and the 8-byte suffix of the memfd
name in a map.  In memfd_fcntl(), we check if the memfd we want is being
sealed (indicating that the tracer has finished writing it), and report
that to user space.

While this should solve many cases, note that there is still a
possibility of race if the process exits even before we are able to read
the memfd from userspace.  A solution will be implemented for that in
the separately in the future (since it requires changes to the tracers,
to use eRPC or something similar).