Fetch CCRID for cases where k8s returns HostAliases (#41653)
### What does this PR do?

The original code for getting CCRIDs doesn't work properly when running under k8s. The call to HostAliases may return "kubelet" as the cloud provider, meaning we don't know which type of CCRID to request. In such cases, try each of them one at a time.

### Motivation

Correctly fetch CCRID when possible. The expectation is that agents running on AWS / Azure / GCP should all contain a CCRID in their host metadata payload.

### Describe how you validated your changes

Manually tested by running in a container in staging.

### Additional Notes