Fixing flaky cert store e2e tests (#41663)
### What does this PR do?
This PR fixes the flaky Windows Certificate Store e2e tests. The tests can fail when restarting the datadogagent service. This is due to the test trying to restart the Agent before it is in the `RUNNING` state. The test also moves some of its setup to the provisioner so that the VM is configured before tests are run.

### Motivation
https://datadoghq.atlassian.net/browse/WINA-1724
https://datadoghq.atlassian.net/browse/WINA-1578

### Describe how you validated your changes
Ran the cert store e2e tests

### Additional Notes