feat: add option to fetch_secret.sh to format secret (#41662)
### What does this PR do?

Adds option to `fetch_secret.sh` to format the secret retrieved from
vault. The option is not required and the default format is table (the
same default that vault has). Therefore, the script should work exactly
the same as before.

### Motivation

The below ticket requires the fetched secret to be in the json format.
The script did not have the ability to specify format.

https://datadoghq.atlassian.net/browse/BARX-1081

### Describe how you validated your changes

Script was ran locally and verified to work.

### Additional Notes