usm, cnm: Add process tags from tagger (#41347)
### What does this PR do?

Get the tags for the Process entity from the tagger and add them to the
connection tags.  If service discovery is enabled, this will include
things like UST (will be de-duplicated in the encoder since system-probe
also gets it directly for USM/CNM) as well as process tags from tracer
metadata.

Without service discovery enabled, this will currently not add any tags.

There are races to be addressed in the future regarding during process startup and for short-lived processes where the workloadmeta may not have the process information, but the other cases should work.

### Motivation

https://datadoghq.atlassian.net/browse/DSCVR-197

### Describe how you validated your changes

Tests added.  Also checked manually with debug prints that the connection check
gets the tags.


Co-authored-by: Yumasi <guillaume.pagnoux@datadoghq.com>