Temporary clone of Bazel's built in repository rules
Copied from https://github.com/bazelbuild/bazel/tree/master/tools/build_defs/repo

This copy brings in some required capabilities to http_archive that are upstreamed
but not yet in a Bazel release. This is preferred over a dependency on all of
Bazel for speed - we won't have to download a gigantic archive to depend on just
a few files.