[BARX-1031] Prevent version regression of "stable" release candidate tags  (#41746)
### What does this PR do?

Only deploy mutable rc image tags (i.e. `7-rc-full`) if the rc tag is the latest tag.

The job `deploy_latest_rc_only` is added as a dependency to jobs that deploy mutable rc image tags. The `deploy_latest_rc_only` job works by listing current tags in git and checking if the current tag (`CI_COMMIT_TAG`) is the latest tag. 

For example, if the current tag is `7.67.0-rc.3` and there is a tag for `7.68.0-rc.1` the job will fail and none of the mutable image tags will be deployed.

### Motivation

https://datadoghq.atlassian.net/browse/BARX-1031

### Describe how you validated your changes

### Additional Notes