dyninst: only run all-probes in debug by default (#41797)
### What does this PR do?

For the integration tests, for the "regular" mode we run all the probes, for all the configs, for all the binaries both as individual and as an aggregate and ensure the results all match. We were also exercising this entire matrix for debug mode. This is not necessary, and we've started to hit timeouts in CI sometimes. Cutting this out should reduce the total execution time by a good 30% or so (maybe more).

One day we'll need to do something more drastic to tame the runtime growth due to combinatorial explosion, but for now, this should do.

### Motivation

We've started to hit timeouts in CI sometimes.

### Describe how you validated your changes

It's all testing.