[EBPF-829] gpu: retry launching Docker workload in e2e test (#41690)
### What does this PR do?

Retries launching e2e workloads if they fail.

### Motivation

There's an error about the workloads not being able to access CUDA devices, which is a source of flakiness in the tests, and the workload is not part of the code under test. If that eliminates these errors we can reduce flakiness in the TestVectorAddProgram test.

### Describe how you validated your changes

Validated locally, will watch CI visibility for flakiness reduction.

### Additional Notes