Add new tags, improve error handling, add func as unique key (#41871)
### What does this PR do?

This improve the `fuzz_infra.py` script to handle ctx.run throws that we encountered in this job:
https://gitlab.ddbuild.io/DataDog/datadog-agent/-/jobs/1175663441

That made the list of fuzzer to be started truncated to the first crashing build.

This also adds the `nvml` tag to build some new targets. (TODO: how can we automate the list of tags to include "everything" ? Now that we can fail a build safely, should we "add them all" ?)

And finally, there's a small improvements on the naming scheme for Go fuzzer: the function name is now part of the package. This creates duplicate binaries, but if a package has  multiple fuzzers, we can now run them properly (i.e: with their own inputs/crash/coverage)

### Motivation
Improve the fuzzing infrastructure error handling and improve coverage

### Describe how you validated your changes
Gitlab run with the build failure not blocking the remaining fuzzers: https://gitlab.ddbuild.io/DataDog/datadog-agent/-/jobs/1175824522

### Additional Notes
N/A