upgrade(fleet): Send updated metadata payload as soon as the experiment starts (#41867)
### What does this PR do?
Adds `DD_INVENTORIES_FIRST_RUN_DELAY=5` to experiment units. This ensures the inventory payloads are send as soon as the Agent Experiment starts (with a small race-preventing delay), eliminating a whole minute of waiting in the UI.

We require waiting for 60s as soon as an Agent starts; but if the Agent is already registered in our DB this delay isn't mandatory. And the Agent must be in our DB to start an upgrade remotely.

### Motivation
More responsive upgrades

### Describe how you validated your changes
Manual QA

### Additional Notes