Add info about agent version + disable cache (#41885)
### What does this PR do?

Make sure git cache is disabled on the setup_agent job. Otherwise it can cause issues when a tag is deleted from upstream, but still present in the cache. And the computed agent version can be wrong
Also display the content of the cache file to make it easier to debug

### Motivation

### Describe how you validated your changes

### Additional Notes