[EBPF] gpu: add nvml state tracker for telemetry (#41823)
### What does this PR do?

Adds a tracker for the state of the NVML library that reports how many times the initialization was tried, and whether the library continues to be unavailable after a period of time.

### Motivation

After https://github.com/DataDog/datadog-agent/pull/41518, both the agent and system-probe GPUm code can continue to execute without an initialized NVML library. With that PR, we can no longer rely on agent check or system-probe initialization errors to monitor for issues with NVML access. We also cannot rely completely on the logs showing NVML errors, as it's hard to monitor which of those errors are temporary and which are not. This PR brings in metrics that we will be able to monitor directly for permanent NVML errors.

### Describe how you validated your changes

Validated locally that metrics are emitted. Unit tests are also included.

### Additional Notes