[EBPF-855] client: fix flaky test (#41953)
### What does this PR do?

Fixes a flaky test introduced in #41708, where a singleton was not being reset correctly. It also correctly renames the test file to check_test.go.

### Motivation

Fix flaky test

### Describe how you validated your changes

Tested by running locally all of the tests in the package in a single run, not individually (this stopped me from seeing the error while developing).

### Additional Notes