dyninst/irgen: deflake TestIrgenMemoryUse (#41971)
### What does this PR do?

The test was somewhat misguided in that it was using the peak heap usage and relying on GC running in a timely manner. Instead, I realized we can use the gctrace to get visibility into the retained heap size after GC concludes. This better captures the intention of the test, and allows for tighter bounds with increased reliability.

### Motivation

We saw a flake on main. I could reproduce by heavily overloading locally using [stress](https://pkg.go.dev/golang.org/x/tools/cmd/stress).


### Describe how you validated your changes

```
2m54s: 2000 runs total, 0 failures
```