[EBPF-851] fix(corechecks/gpu): lazily init gpu device events collector with registration retry logic (#42024)
### What does this PR do?

Introduces a lazy initialization approach for registering devices for event watching in the respective GPU check collectors. Registration is moved at metric collection time, and reattempted up to a max number of times. After the limit is reached, a warning is logged and metric production is skipped for that collector, which however is properly created and never skipped at construction time.

### Motivation

We're hitting sporadic failures when registering a GPU device for events watching (just XID errors so far). The current hypothesis is that this could happen due to a setup/init race when a node is provisioned in autoscaling, and the Nvidia driver is not yet "ready" for registering devices for event watching. This alternative proposal adds resiliency through controlled reattempts at subsequent check runs.

### Describe how you validated your changes

### Additional Notes