[EBPF-846] gpu: add gpu.device_total as a COAT metric (#41803)
### What does this PR do?

Adds `gpu.device_total` as a cross-org agent telemetry metric.

### Motivation

Reliably measure the number of agents with GPU monitoring enabled and the number of devices they're using, for product analytics purposes.

### Describe how you validated your changes

Checked that the value of the metric reports the number of GPU devices in the system, both in normal telemetry and with `agent diagnose show-metadata agent-telemetry`.

### Additional Notes

[Associated ASUP ticket](https://datadoghq.atlassian.net/browse/ASUP-57)