config/remote/service: remove unused abstraction
This `Service` structure is exported but with no users and now no methods.
Furthermore, it embeds a mutex and two fields, but the two fields that it
embeds don't actually use that mutex -- they are never modified.

Instead we hoist the fields up to the `CoreAgentService`, the actual user
of these fields.