Support map[string]string leaf values, more tests
The config uses the schema to ensure that leaf nodes are able to have map values for an individual setting. But this wasn't correctly converting map[string]string because it was only looking for map[string]interface. Use the helper function ToMapStringInterface instead when copying maps in order to handle arbitrary map types.